Update some scenes to use the FSM of gdx ai
Makes the code a little bit longer, but we also help privitize some
features and procedures while providing a consistent way of handling the
UI's states.

Old code was really sloppy with using ints as states.  My only excuse
was because I wrote it in a week.  Seriously, never ever do that for
long term maintable projects.