Pretty up boss combat, add combat notifications, clean up general messaging
It was getting hard keeping track of all the values I was using for
various scene's personal messaging, that I realized that some values
were being reused, which was troublesome for non-direct messaging.
Putting all message constant values in one file will keep things
straight, even if it's not the most kosher way of doing things.